backendfapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
google-generativeai
passlib
argon2-cffi
python-jose
python-dotenv
fastapi-mail
